requirements.txt
mysql-connector-python~=8.2.0
fastapi~=0.104.0
pydantic~=2.4.2
opencv-python~=4.8.1.78
numpy~=1.26.1
requests~=2.31.0
